fix: 변수 이름 변경

+++ backend/src/main/java/com/eumpyo/eum/api/request/PictureAddReq.java
@@ -4,6 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class PictureAddReq {
-    int groupId;
-    int pinId;
+    Long groupId;
+    Long pinId;
 }
